[issue-238] refactor: split grid.py and preferences.py, and the 170-line override writer
Three oversized things the stability fixes have to land in.

ui/grid.py, 2,252 lines and five responsibilities, is 752:

- ui/rom_card.py -- RomItem and its async artwork pipeline: the placeholder,
  the cover fetch, the cartridge composite, the reveal, hover, focus.
- ui/grid_selection.py -- the selection model mapping and the rubber band,
  over a small interface the grid provides (entries, visible entries, the
  cards on screen). The band's rule is a pure function now.
- ui/card_layout.py -- the aspect table, the card sizes, the column packing
  and the cartridge shell, which both the card and the grid measure against.
- The card's menu assembly moves fully into ui/rom_context.py, which already
  built the data-driven half of the same menu. A card widget no longer owns a
  hundred lines of menu policy.

ui/preferences.py had a complete input-capture state machine inside the
dialog: which action is armed, where map-all has got to, and which action has
to let go of a token. It is core/input_capture.py now -- no widget in it -- and
the dialog asks it what happened and renders that. The dialog sits around 10%
coverage; the machine can be tested without a display.

core/retroarch_launcher.py's _write_runtime_override assembled seven concerns
in 170 lines, each block carrying its own comment saying which concern it was.
They are helpers returning dicts (_input_overrides, _bios_overrides,
_shader_overrides, _session_overrides, _av_overrides, _savestate_overrides,
_embed_overrides) merged by a thin writer. The output is byte-identical:
verified by generating the override for three consoles before and after.

Behavior-preserving throughout. 1,711 tests pass (56 new, covering the capture
machine, the band's rule and each override piece -- including the embed block
that used to leak into users' own retroarch.cfg); the app was driven in the
devbox through the grid, a rubber-band selection, a card's context menu, the
input capture with a colliding binding, and a map-all interrupted by Escape.

tests/test_preferences.py becomes tests/test_input_capture.py, following the
code it tests.

Test book: RT-245 to RT-248 added, RT-074 follows its suite file.
